requirements.txt
streamlit>=1.32.0
pydantic>=2.6.0
numpy>=1.24.0
regex>=2023.12.25
scikit-learn>=1.2.2,<1.4
tqdm>=4.66.0
faiss-cpu>=1.7.4
sentence-transformers>=2.5.0
huggingface_hub>=0.22.0
requests>=2.31.0
protobuf==3.20.3
pdfminer.six>=20221105
beautifulsoup4>=4.12.0
lxml>=4.9.0